+++ spec/export_spec.rb
@@ -32,7 +32,7 @@ describe "Export" do
 
     expect($mystring.include?("Compiling src/x.cpp")).to be == true
     expect($mystring.include?("Creating build/test_main_rel_test/liblib3.a")).to be == true
-    expect($mystring.include?("Linking build/rel_test/main.exe")).to be == true
+    expect($mystring.include?("Linking build/rel_test/main"+Bake::Toolchain.outputEnding)).to be == true
     expect($mystring.include?("Rebuilding done.")).to be == true
   end
 
@@ -41,7 +41,7 @@ describe "Export" do
 
     expect($mystring.include?("Compiling src/x.cpp")).to be == false
     expect($mystring.include?("liblib3.a")).to be == false
-    expect($mystring.include?("Linking build/rel_test/main.exe")).to be == false
+    expect($mystring.include?("Linking build/rel_test/main"+Bake::Toolchain.outputEnding)).to be == false
     expect($mystring.include?("Building done.")).to be == true
   end
 
@@ -52,7 +52,7 @@ describe "Export" do
 
     expect($mystring.include?("Compiling src/x.cpp")).to be == false
     expect($mystring.include?("liblib3.a")).to be == false
-    expect($mystring.include?("Linking build/rel_test/main.exe")).to be == true
+    expect($mystring.include?("Linking build/rel_test/main"+Bake::Toolchain.outputEnding)).to be == true
     expect($mystring.include?("Rebuilding done.")).to be == true
   end
   it 'Without file clean' do
@@ -64,7 +64,7 @@ describe "Export" do
 
     expect($mystring.include?("Compiling src/x.cpp")).to be == false
     expect($mystring.include?("liblib3.a")).to be == false
-    expect($mystring.include?("Linking build/rel_test/main.exe")).to be == true
+    expect($mystring.include?("Linking build/rel_test/main"+Bake::Toolchain.outputEnding)).to be == true
     expect($mystring.include?("Building done.")).to be == true
   end
   it 'Without file lib' do
@@ -72,7 +72,7 @@ describe "Export" do
 
     expect($mystring.include?("Compiling src/x.cpp")).to be == false
     expect($mystring.include?("liblib3.a")).to be == false
-    expect($mystring.include?("Linking build/rel_test/main.exe")).to be == false
+    expect($mystring.include?("Linking build/rel_test/main"+Bake::Toolchain.outputEnding)).to be == false
     expect($mystring.include?("Building done.")).to be == true
   end
   it 'Without file lib rebuild' do
@@ -80,7 +80,7 @@ describe "Export" do
 
     expect($mystring.include?("Compiling src/x.cpp")).to be == false
     expect($mystring.include?("liblib3.a")).to be == false
-    expect($mystring.include?("Linking build/rel_test/main.exe")).to be == false
+    expect($mystring.include?("Linking build/rel_test/main"+Bake::Toolchain.outputEnding)).to be == false
     expect($mystring.include?("Rebuilding done.")).to be == true
   end
   it 'Without file main rebuild' do
@@ -88,7 +88,7 @@ describe "Export" do
 
     expect($mystring.include?("Compiling src/x.cpp")).to be == false
     expect($mystring.include?("liblib3.a")).to be == false
-    expect($mystring.include?("Linking build/rel_test/main.exe")).to be == true
+    expect($mystring.include?("Linking build/rel_test/main"+Bake::Toolchain.outputEnding)).to be == true
     expect($mystring.include?("Rebuilding done.")).to be == true
   end
   it 'With file again build' do
@@ -100,7 +100,7 @@ describe "Export" do
 
     expect($mystring.include?("Compiling src/x.cpp")).to be == true
     expect($mystring.include?("Creating build/test_main_rel_test/liblib3.a")).to be == true
-    expect($mystring.include?("Linking build/rel_test/main.exe")).to be == true
+    expect($mystring.include?("Linking build/rel_test/main"+Bake::Toolchain.outputEnding)).to be == true
     expect($mystring.include?("Building done.")).to be == true
   end