+++ spec/multiple_root_spec.rb
@@ -13,55 +13,55 @@ module Bake
 describe "Multiple root" do
 
   it 'single root' do
-    expect(File.exists?("spec/testdata/root1/main/build/test/main.exe")).to be == false
+    expect(File.exists?("spec/testdata/root1/main/build/test/main"+Bake::Toolchain.outputEnding)).to be == false
     Bake.startBake("root1/main", ["test"])
     expect(ExitHelper.exit_code).to be > 0
     expect($mystring.split("Project.meta not found").length).to be == 2
   end
 
   it 'both roots' do
-    expect(File.exists?("spec/testdata/root1/main/build/test/main.exe")).to be == false
+    expect(File.exists?("spec/testdata/root1/main/build/test/main"+Bake::Toolchain.outputEnding)).to be == false
     Bake.startBake("root1/main", ["test", "-w", "spec/testdata/root1", "-w", "spec/testdata/root2", "--adapt", "nols"])
-    expect(File.exists?("spec/testdata/root1/main/build/test/main.exe")).to be == true
+    expect(File.exists?("spec/testdata/root1/main/build/test/main"+Bake::Toolchain.outputEnding)).to be == true
   end
 
   it 'wrong root' do
-    expect(File.exists?("spec/testdata/root1/main/build/test/main.exe")).to be == false
+    expect(File.exists?("spec/testdata/root1/main/build/test/main"+Bake::Toolchain.outputEnding)).to be == false
     expect { Bake.startBake("root1/main", ["test", "-w", "spec/testdata/root1", "-w", "spec/testdata/root2/lib3", "--adapt", "nols"]) }.to raise_error(SystemExit)
     expect(ExitHelper.exit_code).to be > 0
     expect($mystring.split("lib3 does not exist").length).to be == 2
   end
 
   it 'forgotten root' do
-    expect(File.exists?("spec/testdata/root1/main/build/test/main.exe")).to be == false
+    expect(File.exists?("spec/testdata/root1/main/build/test/main"+Bake::Toolchain.outputEnding)).to be == false
     Bake.startBake("root1/main", ["test", "-w", "spec/testdata/root1", "--adapt", "nols"])
     expect(ExitHelper.exit_code).to be > 0
     expect($mystring.split("Error: lib2/Project.meta not found").length).to be == 2
   end
 
   it 'invalid root' do
-    expect(File.exists?("spec/testdata/root1/main/build/test/main.exe")).to be == false
+    expect(File.exists?("spec/testdata/root1/main/build/test/main"+Bake::Toolchain.outputEnding)).to be == false
     expect { Bake.startBake("root1/main", ["test", "-w", "spec/testdata/GIBTS_DOCH_GAR_NICHT", "--adapt", "nols"]) }.to raise_error(SystemExit)
     expect($mystring.split("Error: Directory spec/testdata/GIBTS_DOCH_GAR_NICHT does not exist").length).to be == 2
   end
 
   it 'auto root overwritten wrong' do
-    expect(File.exists?("spec/testdata/root1/mainAutoRoot/build/test/mainAutoRoot.exe")).to be == false
+    expect(File.exists?("spec/testdata/root1/mainAutoRoot/build/test/mainAutoRoot"+Bake::Toolchain.outputEnding)).to be == false
     Bake.startBake("root1/mainAutoRoot", ["test", "-w", "spec/testdata/root1", "--adapt", "nols"])
     expect($mystring.include?("Error: lib2")).to be == true
-    expect(File.exists?("spec/testdata/root1/mainAutoRoot/build/test/mainAutoRoot.exe")).to be == false
+    expect(File.exists?("spec/testdata/root1/mainAutoRoot/build/test/mainAutoRoot"+Bake::Toolchain.outputEnding)).to be == false
   end
 
   it 'auto root overwritten right' do
-    expect(File.exists?("spec/testdata/root1/mainAutoRoot/build/test/mainAutoRoot.exe")).to be == false
+    expect(File.exists?("spec/testdata/root1/mainAutoRoot/build/test/mainAutoRoot"+Bake::Toolchain.outputEnding)).to be == false
     Bake.startBake("root1/mainAutoRoot", ["test", "-w", "spec/testdata/root1", "-w", "spec/testdata/root2", "--adapt", "nols"])
-    expect(File.exists?("spec/testdata/root1/mainAutoRoot/build/test/mainAutoRoot.exe")).to be == true
+    expect(File.exists?("spec/testdata/root1/mainAutoRoot/build/test/mainAutoRoot"+Bake::Toolchain.outputEnding)).to be == true
   end
 
   it 'auto root' do
-    expect(File.exists?("spec/testdata/root1/mainAutoRoot/build/test/mainAutoRoot.exe")).to be == false
+    expect(File.exists?("spec/testdata/root1/mainAutoRoot/build/test/mainAutoRoot"+Bake::Toolchain.outputEnding)).to be == false
     Bake.startBake("root1/mainAutoRoot", ["test", "--adapt", "nols"])
-    expect(File.exists?("spec/testdata/root1/mainAutoRoot/build/test/mainAutoRoot.exe")).to be == true
+    expect(File.exists?("spec/testdata/root1/mainAutoRoot/build/test/mainAutoRoot"+Bake::Toolchain.outputEnding)).to be == true
   end
 
   it 'roots.bake' do